Rename source files to UpperCamelCase; add CMake build + BUILD.md
File naming: rename all Core, Driver, and public-header source files from
snake_case to UpperCamelCase (HardwareSnapshot.cpp, Vdp1Parser.h, SeHost.h, …),
matching the FrontEnd (which was already UpperCamelCase) and the project's
UpperCamelCase functions/types. The public C-ABI *identifiers* stay snake_case
(se_create, se_result) — that's the stable C interface; only the file names
change. All #include directives and the .vcxproj/.filters project files are
updated to match; ARCHITECTURE.md §12 documents the file-name rule.

Build system: add a CMake build (SaturnExplorer/CMakeLists.txt) so the project
builds without Visual Studio and can *generate* a VS solution
(cmake -B build -G "Visual Studio 17 2022" -A x64). Core and the savestate
driver are portable and build on Windows/macOS/Linux; the Win32/D3D11 FrontEnd
is built only on WIN32 (skipped elsewhere with a message). BUILD.md documents
both the CMake and the checked-in-.sln paths; .gitignore excludes CMake/VS
build trees and stray render output.

Verified: CMake configures and builds Core + driver on Linux; the core+driver
compile clean under clang with the new names and render Battle3 byte-identical;
App.cpp compiles clean against ImGui.
