+++ SaturnExplorer/.gitignore
@@ -1,0 +1,21 @@
+# CMake out-of-source build trees + generated files
+build/
+out/
+CMakeCache.txt
+CMakeFiles/
+cmake_install.cmake
+CTestTestfile.cmake
+
+# Visual Studio / MSBuild
+.vs/
+*.user
+*.suo
+*.vcxproj.user
+[Dd]ebug/
+[Rr]elease/
+x64/
+Win32/
+Libs/
+
+# Stray render / test output (harnesses write PNGs; none are tracked here)
+*.png
