+++ SaturnExplorer/include/saturnexplorer/SaturnExplorer.h
@@ -1,14 +1,14 @@
 /* Saturn Explorer — umbrella include.
  *
  * Include this to get the full public API. Drivers typically need only
- * se_data_source.h; hosts need se_host.h. See ARCHITECTURE.md.
+ * SeDataSource.h; hosts need SeHost.h. See ARCHITECTURE.md.
  */
 #ifndef SATURNEXPLORER_H
 #define SATURNEXPLORER_H
 
-#include "se_abi.h"
-#include "se_types.h"
-#include "se_data_source.h"
-#include "se_host.h"
+#include "SeAbi.h"
+#include "SeTypes.h"
+#include "SeDataSource.h"
+#include "SeHost.h"
 
 #endif /* SATURNEXPLORER_H */
